move haplotype to separate classes
